プロフィールページのコンポーネント構成を変更
